fix(metrics): fix the init issue from the previous refactor of async metrics collection
