+++ internal/operators/v1/metrics/metrics.go
@@ -1,4 +1,4 @@
-package supportfiles
+package metrics
 
 import (
 	"context"
