+++ internal/runtime/node.go
@@ -92,7 +92,6 @@ func initNodeIdentities() error {
 	definition.LocalTimeFixedZone = time.FixedZone("", definition.LocalTimeZoneSeconds)
 	cubecos.SyncTunings()
 	cubecos.SyncSupportFiles()
-	cubecos.SyncDataCenterMetricsSummary()
 
 	return nil
 }
